requirements.txt
pandas>=2.0.0
numpy>=1.24.0
requests>=2.28.0
scikit-learn>=1.2.0
plotly>=5.14.0
streamlit>=1.30.0
pydantic>=2.0.0
transformers>=4.30.0
torch>=2.0.0
feedparser>=6.0.0
beautifulsoup4>=4.12.0
newspaper3k>=0.2.8
geopy>=2.3.0
rapidfuzz>=3.0.0
python-dateutil>=2.8.0
lxml>=4.9.0
